# docs/waystone-sample-api-swagger.json
{
  "x-generator": "NSwag v13.18.0.0 (NJsonSchema v10.8.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "Waystone.Sample.Api",
    "description": "A sample API built with Waystone.Common.Api",
    "version": "v1"
  },
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List a page of products.",
        "operationId": "Products_List",
        "parameters": [
          {
            "name": "Limit",
            "in": "query",
            "description": "The maximum number of records to return;",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 10
            },
            "x-position": 1
          },
          {
            "name": "Cursor",
            "in": "query",
            "description": "The offset (from 0) of the first record to return;",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 0
            },
            "x-position": 2
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "The correlation id to be associated with the request. One will be generated by the API if not provided.",
            "schema": {
              "type": "string",
              "format": "guid",
              "example": "72b264f5-d6e9-4e74-a6ce-d398e2c8421a"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnknownProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "A paginated response of products.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfProductDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Adds a new product.",
        "operationId": "Products_Add",
        "parameters": [
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "The correlation id to be associated with the request. One will be generated by the API if not provided.",
            "schema": {
              "type": "string",
              "format": "guid",
              "example": "72b264f5-d6e9-4e74-a6ce-d398e2c8421a"
            }
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCommand"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnknownProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "The created product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Gets a product using it's ID.",
        "operationId": "Products_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the product.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "The correlation id to be associated with the request. One will be generated by the API if not provided.",
            "schema": {
              "type": "string",
              "format": "guid",
              "example": "72b264f5-d6e9-4e74-a6ce-d398e2c8421a"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnknownProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "The product corresponding to the provided ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "The correlation id to be associated with the request. One will be generated by the API if not provided.",
            "schema": {
              "type": "string",
              "format": "guid",
              "example": "72b264f5-d6e9-4e74-a6ce-d398e2c8421a"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnknownProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NotFoundProblemDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StatusCodeProblemDetails"
          },
          {
            "type": "object",
            "description": "Not found problem details.",
            "additionalProperties": {
              "nullable": true
            }
          }
        ]
      },
      "StatusCodeProblemDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "UnknownProblemDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StatusCodeProblemDetails"
          },
          {
            "type": "object",
            "description": "Internal server error problem details.",
            "additionalProperties": {
              "nullable": true
            }
          }
        ]
      },
      "PaginatedResponseOfProductDto": {
        "type": "object",
        "description": "Represents a page of records.",
        "additionalProperties": false,
        "properties": {
          "links": {
            "description": "The previous, current, and next page Links.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Links"
              }
            ]
          },
          "total": {
            "type": "integer",
            "description": "The total number of records available.",
            "format": "int32",
            "example": 100
          },
          "results": {
            "type": "array",
            "description": "The current page of records.",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          }
        }
      },
      "Links": {
        "type": "object",
        "description": "The links used to navigate between pages in a paginated collection.",
        "additionalProperties": false,
        "properties": {
          "self": {
            "type": "string",
            "description": "The URI link to the current page of results.",
            "format": "uri",
            "nullable": true
          },
          "next": {
            "type": "string",
            "description": "The URI link to the next page of results if it exists.",
            "format": "uri",
            "nullable": true
          },
          "previous": {
            "type": "string",
            "description": "The URI link to the previous page of results if it exists.",
            "format": "uri",
            "nullable": true
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "amountExcludingTax": {
            "type": "number",
            "format": "decimal"
          },
          "tax": {
            "type": "number",
            "format": "decimal"
          },
          "total": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
      "CreateProductCommand": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "amountExcludingTax": {
            "type": "number",
            "format": "decimal"
          },
          "taxPercentage": {
            "type": "number",
            "format": "decimal"
          },
          "discountPercentage": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      }
    }
  }
}